# openapi-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "EvoNEST API",
    "version": "2.0.0",
    "description": "\n# EvoNEST General Documentation\n\n**The EvoNEST User Guide and the EvoNEST Developer Guide are found here**: https://daniele-liprandi.github.io/EvoNEST-backbone/\n\n# EvoNEST API Documentation\n\nThe EvoNEST API allows users to interact with the MongoDB databases generated in their NEST, plus it gives access to utility functions useful in ecological, evolutionary and biological research.\n\n## Features\n\n- **Samples Management**: Create, update, and retrieve biological samples with taxonomic information\n- **Traits Measurement**: Record and analyze trait measurements with statistical calculations\n- **Experiments**: Manage measurements, documents and raw data\n- **File Storage**: Upload and manage research files with metadata\n- **User Management**: Handle user accounts and permissions\n- **Utilities**: Geocoding, species image search, and data processing tools\n\n## Authentication\n\nMost endpoints require proper authentication and database access permissions.\n\n## Data Format\n\nAll endpoints return JSON data unless otherwise specified. Dates are in ISO 8601 format.\n      ",
    "contact": {
      "name": "Daniele Liprandi",
      "email": "daniele.liprandi@gmail.com",
      "url": "https://github.com/daniele-liprandi/EvoNEST-backbone"
    },
    "license": {
      "name": "GNU Affero General Public License v3.0",
...
    },
...
    "x-build-context": "GitHub Actions"
...
  "servers": [
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "sessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    },
    {
      "sessionAuth": []
    }
  ],
  "paths": {
...
    "/api/user/database": {
      "get": {
        "summary": "Get user's database access information",
        "description": "Retrieve list of databases the user has access to and their currently active database",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "SessionAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Database information retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "databases": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "description": "List of databases user has access to"
                    },
                    "activeDatabase": {
                      "type": "string",
                      "description": "Currently active database for the user"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "post": {
        "summary": "Set active database for user",
        "description": "Switch the user's active database to a different one they have access to",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "SessionAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "database"
                ],
                "properties": {
                  "database": {
                    "type": "string",
                    "description": "Name of the database to set as active"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Active database updated successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Database not authorized for user"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
  }
}